Update supported platforms tests
* Add Molecule testing for supported platforms

The Molecule tests correct a ZooKeeper cluster consisting of three nodes running
within Docker containers. Each Docker container is a different OS to test the platforms
that this role supports.

Upgrade to Molecule 3:

- playbook.yml was deprecated as has been moved to converge.yml
- linting is run using a shell command for yamllint and ansible-lint

* playbook.yml renamed to converge.yml for Molecule 3 upgrade

* ZooKeeper service exit status code set to 143

Java processes exit with a status code of 143 when stopped. This is a non-zero exit code so
can appear as a failure within systemd unless explicitly configured as a success.

* Add Molecule verifier playbook for testing

The verify.yml playbook is run during the Verify step during Molecule testing
and asserts that ZooKeeper is installed and running on all servers as expected.

* Add GitHub Action to run Molecule tests

- Update role name in converge.yml and remove redundant python Molecule tests
- Use command syntax to install dependencies and run Molecule test scenario
- Checkout source to the path "sleighzy.zookeeper" to match the role name