Upgrade to Apache ZooKeeper 3.7.0 (#17)
* Upgrade to ZooKeeper 3.7.0

Upgrade to the Apache ZooKeeper 3.7.0 release

* Breaking change with kernel versions and systemd

Some kernel versions have made breaking changes to the attributes
returned in systemd service calls that Ansible uses to determine what
state the service is in. Ansible 2.9.16 and 2.10.4 are the minimum
required versions as a workaround was put in place to use different
commands to retrieve the service status.

* Molecule Docker driver breaking change

A breaking change in Molecule 3.1.1 removed the Docker driver.
This must be installed separately using the pip molecule-docker
module for the Molecule tests to run.