Add namespace and role name to Galaxy info
The namespace and role name are required as Ansible Galaxy v3.0 does not
alter the GitHub repository name when importing the role so this needs
to be done explicitely.

This resolves the issue where this is tested for by Molecule which would
have produced the below error.

ERROR    Computed fully qualified role name of zookeeper does not follow current galaxy requirements.