Remove Rocky Linux and use Redhat 9 due to breaking Python changes
The ansible-core 2.17 release has dropped support for Python 3.6, which is the default in Rocky Linux (as it is an EL8 platform).
This leads to the "future feature annotations is not defined" error when gathering facts or installing packages.
Installing Python 3.9 also does not work as the issue lies in /usr/libexec/platform-python which is version 3.6,
and is the "system" python which is used for installing modules using dnf or yum.

To resolve this issue the testing containers have been updated to use Redhat 9.